improve: cmdSelector
